pyproject.toml
[tool.poetry]
name = "brewblox-plaato"
version = "0.1.0"
description = "Brewblox service for the Plaato airlock"
authors = ["BrewPi <development@brewpi.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4"
brewblox-service = "^3.3.2"

[tool.poetry.group.dev.dependencies]
aresponses = "*"
pytest-cov = "*"
pytest-mock = "*"
pytest-aiohttp = "*"
flake8 = "*"
autopep8 = "*"
pytest = "*"
flake8-quotes = "*"
invoke = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
